fix(vllm): move sqlite repository into db management
